Move analysis from fragment, add LiveAnalysisRunner, remove LiveAnalysis
LiveAnalysis can be handled by Default/CustomAnalysis
move make_value_dict and make_coordinate_dict to utils
