+++ ndscan/experiment/fragment.py
@@ -4,7 +4,7 @@ from copy import deepcopy
 import logging
 from typing import Any, Dict, List, Iterable, Type, Tuple, Optional
 
-from .default_analysis import DefaultAnalysis, LiveAnalysis, ResultPrefixAnalysisWrapper
+from .default_analysis import DefaultAnalysis, ResultPrefixAnalysisWrapper
 from .parameters import ParamHandle, ParamStore
 from .result_channels import ResultChannel, FloatChannel
 from .utils import is_kernel, path_matches_spec
@@ -707,70 +707,8 @@ class ExpFragment(Fragment):
         """
         return []
 
-    def get_live_analyses(self) -> List[LiveAnalysis]:
+    def get_live_analyses(self) -> List[DefaultAnalysis]:
         return []
-
-    def register_live_analyses(
-        self,
-        axes=[],
-        coordinate_sinks=[],
-        scan_result_sinks: dict = {},
-    ):
-        """Filter analyses for specific scan and store"""
-        analyses = self.get_live_analyses()
-        self._analyses = filter_live_analyses(analyses, axes)
-        self._axes = axes
-        self._coordinate_sinks = coordinate_sinks
-        self._scan_result_sinks = scan_result_sinks
-
-    def _make_coordinate_dict(self):
-        """Copy paste from entry_point/TopLevelRunner._make_coordinate_dict
-        (ran into an import loop I didn't have the brain power to fix)"""
-        return OrderedDict(
-            ((a.param_schema["fqn"], a.path), s.get_all())
-            for a, s in zip(self._axes, self._coordinate_sinks)
-        )
-
-    def _make_value_dict(self):
-        """Copy paste from entry_point/TopLevelRunner._make_value_dict
-        (ran into an import loop I didn't have the brain power to fix)"""
-        return {c: s.get_all() for c, s in self._scan_result_sinks.items()}
-
-    @rpc(flags={"async"})
-    def live_analyses(self):
-        """Minimal implementation of entry_point/TopLevelRunner.analyze
-        Purely for data processing, does not handle annotations (yet!)"""
-        coordinates = self._make_coordinate_dict()
-        values = self._make_value_dict()
-        for a in self._analyses:
-            a.execute(coordinates, values)
-
-
-def filter_live_analyses(analyses: List[LiveAnalysis], axes) -> List[LiveAnalysis]:
-    """Copy of scan_runner.filter_default_analyses
-    (ran into an import loop I didn't have the brain power to fix)"""
-    ax = list(axes)  # Don't exhaust an arbitrary iterable.
-    result = []
-    for a in analyses:
-        if not isinstance(a, LiveAnalysis):
-            raise ValueError(
-                f"Unexpected get_default_analyses() return value for {a}: "
-                "Expected list of ndscan.experiment.DefaultAnalysis instances, got "
-                f"element of type '{a}'"
-            )
-        if match_default_analysis(a, ax):
-            result.append(a)
-    return result
-
-
-def match_default_analysis(analysis: LiveAnalysis, axes) -> bool:
-    """Copy of scan_runner.filter_default_analyses
-    (ran into an import loop I didn't have the brain power to fix)
-    """
-    stores = set(a.param_store for a in axes)
-    assert None not in stores, "Can only match analyses after stores have been created"
-    return set(a._store for a in analysis.required_axes()) == stores
-
 
 def _skip_common_prefix(target: list, reference: list) -> list:
     i = 0
