+++ ndscan/experiment/scan_runner.py
@@ -16,7 +16,7 @@ from .fragment import ExpFragment, TransitoryError, RestartKernelTransitoryError
 from .parameters import ParamStore, type_string_to_param
 from .result_channels import ResultChannel, ResultSink
 from .scan_generator import generate_points, ScanGenerator, ScanOptions
-from .utils import is_kernel
+from .utils import is_kernel, make_coordinate_dict, make_value_dict
 
 __all__ = [
     "ScanAxis",
@@ -94,7 +94,7 @@ class ScanRunner(HasEnvironment):
         self.setattr_device("scheduler")
 
     def run(
-        self, fragment: ExpFragment, spec: ScanSpec, axis_sinks: List[ResultSink]
+        self, fragment: ExpFragment, spec: ScanSpec, axis_sinks: List[ResultSink], scan_result_sinks: dict,
     ) -> None:
         """Run a scan of the given fragment, with axes as specified.
 
@@ -111,6 +111,9 @@ class ScanRunner(HasEnvironment):
             self._run_scan_on_core_device
             if is_kernel(fragment.run_once)
             else self._run_scan_on_host
+        )
+        self.live_analysis_runner = LiveAnalysisRunner(
+            fragment, spec.axes, axis_sinks, scan_result_sinks
         )
         run_impl(fragment, points, spec.axes, axis_sinks)
 
@@ -138,7 +141,7 @@ class ScanRunner(HasEnvironment):
                             sink.push(value)
                         fragment.device_setup()
                         fragment.run_once()
-                        fragment.live_analyses()
+                        self.live_analysis_runner.live_analyses()
                         if self.scheduler.check_pause():
                             break
                 finally:
@@ -265,7 +268,7 @@ class ScanRunner(HasEnvironment):
             try:
                 self._fragment.device_setup()
                 self._fragment.run_once()
-                self._fragment.live_analyses()
+                self.live_analysis_runner.live_analyses()
                 break
             except RTIOUnderflow:
                 # For the first two underflows per point, just print a warning and carry
@@ -416,6 +419,30 @@ def filter_default_analyses(
         if match_default_analysis(analysis, ax):
             result.append(analysis)
     return result
+
+
+class LiveAnalysisRunner:
+    def __init__(
+        self,
+        fragment,
+        axes: List[ScanAxis] = [],
+        coordinate_sinks: List[ResultSink] = [],
+        scan_result_sinks: dict = {},
+    ):
+        """Filter analyses for specific scan and store"""
+        self._analyses = filter_default_analyses(fragment.get_live_analyses(), axes)
+        self._axes = axes
+        self._coordinate_sinks = coordinate_sinks
+        self._scan_result_sinks = scan_result_sinks
+
+    @rpc(flags={"async"})
+    def live_analyses(self):
+        """Minimal implementation of entry_point/TopLevelRunner.analyze
+        Purely for data processing, does not handle annotations (yet!)"""
+        coordinates = make_coordinate_dict(self._axes, self._coordinate_sinks)
+        values = make_value_dict(self._scan_result_sinks)
+        for a in self._analyses:
+            a.execute(coordinates, values)
 
 
 def describe_scan(
